Fixed environment variable issues

+++ tests/system/testDriver/src/com/bbn/rpki/test/objects/TestbedConfig.java
@@ -66,8 +66,13 @@ public class TestbedConfig implements Constants {
 	public TestbedConfig(String iniFile) {
 		File objectsDir = new File(OBJECT_PATH);
 		//if the directory was not created right now, delete the directories within it
-		if (!objectsDir.mkdir()) {
+		if (!objectsDir.mkdir() && objectsDir.exists()) {
 			Util.deleteDirectories(objectsDir.listFiles());
+		} else if (!objectsDir.exists()) {
+			System.err.println("Could not create directory: " + objectsDir.getAbsolutePath());
+			System.err.println("Source dir: " + Constants.srcDir);
+			System.err.println("Build dir: " + Constants.buildDir);
+			System.err.println("objects dir: " + Constants.OBJECT_PATH);
 		}
 		try {
 			wini = new Wini(new StringReader(iniFile));
